refactor: :recycle: Separated leaflet marker and observables to leaflet.ts
