+++ index.html
@@ -8,7 +8,7 @@
   </head>
 
   <body>
-    <div class="grid-container grid grid-rows-2 lg:grid-cols-large">
+    <div class="grid-container grid grid-rows-2 lg:grid-rows-none lg:grid-cols-large">
       <div class="map h-map-half z-0
       lg:order-2 lg:h-screen
       " id="map"></div>
